feat: focused todo workspace, per-user assistant model, direct calendar booking
Ships the Sep 12-13 working tree so CI deploys it.

Mac companion: runner-theme shell, Tasks table, People and source pages, and
the rebuilt todo workspace (read plus at most two action cards beside a
runner-next style chat pane with live tool steps and queued sends).

Server: users.assistant_model set from the web Settings page and pinned on
every provider call for that user through Maraithon.LLM.UserModel (bound at
chat runs, the chief of staff agent, and background jobs). Calendar create,
update, and cancel actions now execute as soon as they are prepared instead
of waiting for a confirmation tap; messages to other people still get a
review card. Prompt and tool descriptions updated to match.

Also includes the mobile app work, docs, and specs accumulated on this
branch, and records that deploys run only in CI under the deployer's
workload identity.

+++ lib/maraithon/runtime/coordination/fair_scheduler.ex
@@ -97,8 +97,8 @@ defmodule Maraithon.Runtime.Coordination.FairScheduler do
         Repo,
         """
         UPDATE public.runtime_tenant_fairness
-        SET max_concurrency = $2, rate_per_minute = $3, burst = $4,
-            available_microunits = LEAST(available_microunits, $4::bigint * #{@microunits}),
+        SET max_concurrency = $2, rate_per_minute = $3, burst = $4::integer,
+            available_microunits = LEAST(available_microunits, ($4::integer)::bigint * #{@microunits}),
             updated_at = timezone('UTC', clock_timestamp())
         WHERE tenant_key = $1
         """,
